requirements.txt
numpy==1.23.2
opencv-python==4.6.0.66
Pillow==9.2.0
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.1
torch==1.12.1+cu116
torchvision==0.13.1+cu116
tqdm==4.64.1
